Pass File object instead of bufferArray
Since the buffer array was taking > 15 seconds on a fast PC to convert from an
image array to a single dimension Java
byte array, it dawned on me that the processing did not require the image array, only an image that could be loaded.  So while we load the image twice it is much cheaper than converting the loaded image to a Java byte array and there is a significant memory savings.