Fix property array size update
So change events will only fire on arrays if the size of the array changes.  Since we are replacing one currentFolder in the array with another from the binding it appears there is no array size change between process ticks.  We need to "tickle" the array into thinking it was resized.  We do this by clearing the array first and then in a defer (which is next event loop) it will detect a size change and thus rebind the value.